requirements.txt
defusedxml==0.5.0
Django==1.11.8
django-cors-headers==2.2.0
djangorestframework==3.7.7
djangorestframework-jsonapi==2.4.0
djangorestframework-xml==1.3.0
inflection==0.3.1
mysqlclient==1.3.4
pytz==2018.3
six==1.11.0
virtualenv==15.1.0
